Split and remove diagnostic crate
