+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
-members = ["dolang", "dolang-private-build", "dolang-private-bundle", "dolang-private-test", "dolang-private-regression", "dolang-shell-vfs", "dolang-bytecode", "dolang-compile", "dolang-lsp", "dolang-runtime", "dolang-shell-core", "dolang-shell-modules", "dolang-shell", "dolang-private-util", "dolang-ext-shell", "dolang-ext-http", "dolang-ext-url", "dolang-ext-zip", "dolang-ext-sqlite", "dolang-ext-regex", "dolang-ext-glob", "dolang-private-highlight", "dolang-ext-json", "dolang-ext-compile", "dolang-ext-diagnostic", "dolang-ext-load", "dolang-ext-xml", "dolang-ext-yaml", "dolang-ext-toml", "dolang-ext-digest", "dolang-ext-base64", "dolang-ext-progress", "dolang-ext-rand", "dolang-ext-patch", "dolang-private-doc", "dolang-rpc", "fuzz"]
-default-members = ["dolang", "dolang-private-regression", "dolang-shell-vfs", "dolang-bytecode", "dolang-compile", "dolang-lsp", "dolang-runtime", "dolang-shell-core", "dolang-shell-modules", "dolang-shell", "dolang-private-util", "dolang-ext-shell", "dolang-ext-http", "dolang-ext-url", "dolang-ext-zip", "dolang-ext-sqlite", "dolang-ext-regex", "dolang-ext-glob", "dolang-private-highlight", "dolang-ext-json", "dolang-ext-compile", "dolang-ext-diagnostic", "dolang-ext-load", "dolang-ext-xml", "dolang-ext-yaml", "dolang-ext-toml", "dolang-ext-digest", "dolang-ext-base64", "dolang-ext-progress", "dolang-ext-rand", "dolang-ext-patch", "dolang-private-doc", "dolang-rpc"]
+members = ["dolang", "dolang-private-build", "dolang-private-bundle", "dolang-private-test", "dolang-private-regression", "dolang-shell-vfs", "dolang-bytecode", "dolang-compile", "dolang-lsp", "dolang-runtime", "dolang-shell-core", "dolang-shell-modules", "dolang-shell", "dolang-private-util", "dolang-ext-shell", "dolang-ext-http", "dolang-ext-url", "dolang-ext-zip", "dolang-ext-sqlite", "dolang-ext-regex", "dolang-ext-glob", "dolang-private-highlight", "dolang-ext-json", "dolang-ext-compile", "dolang-ext-load", "dolang-ext-xml", "dolang-ext-yaml", "dolang-ext-toml", "dolang-ext-digest", "dolang-ext-base64", "dolang-ext-progress", "dolang-ext-rand", "dolang-ext-patch", "dolang-private-doc", "dolang-rpc", "fuzz"]
+default-members = ["dolang", "dolang-private-regression", "dolang-shell-vfs", "dolang-bytecode", "dolang-compile", "dolang-lsp", "dolang-runtime", "dolang-shell-core", "dolang-shell-modules", "dolang-shell", "dolang-private-util", "dolang-ext-shell", "dolang-ext-http", "dolang-ext-url", "dolang-ext-zip", "dolang-ext-sqlite", "dolang-ext-regex", "dolang-ext-glob", "dolang-private-highlight", "dolang-ext-json", "dolang-ext-compile", "dolang-ext-load", "dolang-ext-xml", "dolang-ext-yaml", "dolang-ext-toml", "dolang-ext-digest", "dolang-ext-base64", "dolang-ext-progress", "dolang-ext-rand", "dolang-ext-patch", "dolang-private-doc", "dolang-rpc"]
 resolver = "3"
 
 [workspace.package]
