--- dolang-ext-diagnostic/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "dolang-ext-diagnostic"
-version = "0.1.0"
-description = "Diagnostic rendering extension for the Do language."
-edition = "2024"
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[lib]
-doctest = false
-
-[dependencies]
-annotate-snippets = "0.12.10"
-anstyle = "1.0.14"
-console = "0.16.2"
-dolang = { path = "../dolang", version = "0.1.0" }
-dolang-ext-compile = { path = "../dolang-ext-compile", version = "0.1.0" }
