--- dolang-ext-diagnostic/src/extension.rs
@@ -1,46 +1,0 @@
-use std::{
-    error,
-    fmt::{self, Debug, Display, Formatter},
-};
-
-use dolang::{
-    compile::Compiler,
-    extension,
-    extension::{Extension, Version},
-    runtime::vm::Builder,
-};
-
-pub struct DiagnosticExt;
-
-#[derive(Debug)]
-pub enum Infallible {}
-
-impl Display for Infallible {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        Debug::fmt(self, f)
-    }
-}
-
-impl error::Error for Infallible {}
-
-impl Extension for DiagnosticExt {
-    type Error = Infallible;
-    const NAME: &str = "dolang-diagnostic";
-    const VERSION: Version = Version {
-        major: 0,
-        minor: 1,
-        patch: 0,
-    };
-    const DESCRIPTION: &str = "Do Diagnostic Extension";
-
-    fn apply_compiler(&self, _compiler: &mut Compiler) -> Result<(), Infallible> {
-        Ok(())
-    }
-
-    fn apply_vm<'v>(&self, builder: &mut Builder<'v>) -> Result<(), Infallible> {
-        crate::diagnostic::configure(builder);
-        Ok(())
-    }
-}
-
-extension!(DiagnosticExt);
