+++ dolang-ext-shell/Cargo.toml
@@ -22,6 +22,8 @@ bstr = "1.12.1"
 typed-path.workspace = true
 shlex = "2.0"
 time = { version = "0.3", features = ["formatting", "parsing"] }
+anstyle = "1.0.14"
+console = "0.16.2"
 
 [dependencies.tokio]
 version = "1.48.0"
@@ -30,6 +32,11 @@ features = ["macros", "rt", "fs", "process", "io-util", "sync", "io-std", "time"
 [dependencies.dolang]
 path = "../dolang"
 version = "0.1.0"
+
+[dependencies.dolang-ext-compile]
+path = "../dolang-ext-compile"
+version = "0.1.0"
+features = ["diagnostic-rendering"]
 
 [dependencies.dolang-shell-vfs]
 path = "../dolang-shell-vfs"
