+++ dolang-shell-core/Cargo.toml
@@ -40,10 +40,6 @@ version = "0.1.0"
 path = "../dolang-ext-compile"
 version = "0.1.0"
 
-[dependencies.dolang-ext-diagnostic]
-path = "../dolang-ext-diagnostic"
-version = "0.1.0"
-
 [dependencies.dolang-ext-load]
 path = "../dolang-ext-load"
 version = "0.1.0"
