+++ test/diagnostic.dol
@@ -1,45 +1,45 @@
 import
   compile
-  diagnostic
   strand
+  term
   test:
     - assert
     - assert_eq
-    - assert_ne
     - test
 
 #[test]
-def compile_with_and_without_color()
+def compile_diagnostic()
   let result = compile.compile "bad.dol" "let =\n"
   let diag = result.diagnostics[0]
 
-  let plain = diagnostic.render_compile_diag $diag color: :never:
-  assert (plain.contains "error:")
-  assert (plain.contains "bad.dol")
-
-  let color = diagnostic.render_compile_diag $diag color: :always:
-  assert_ne $color $plain
+  let rendered = diag.render()
+  assert_eq (type rendered) $term.Text
+  assert ((str rendered).contains "expected at least one item")
+  assert ((str rendered).contains "bad.dol")
+  assert (!(str(rendered).ends_with("\n")))
 
 #[test]
 def error()
   let plain = try
     throw "boom"
   catch e
-    diagnostic.render_error $e color: :never:
-  assert (plain.contains "boom")
+    term.render_error $e
+  assert_eq (type plain) $term.Text
+  assert ((str plain).contains "boom")
+  assert (!(str(plain).ends_with("\n")))
 
 #[test]
 def error_explicit_backtrace()
   let color = try
     throw "boom"
   catch e
-    diagnostic.render_error $e backtrace: (strand.error_backtrace()) color: :always:
-  assert (color.contains "boom")
+    term.render_error $e backtrace: (strand.error_backtrace())
+  assert ((str color).contains "boom")
 
 #[test]
 def error_no_backtrace()
   let err = try
-    diagnostic.render_error "boom"
+    term.render_error "boom"
   catch e
     e
   assert_eq (str err) "state error: no active handled exception"
@@ -47,7 +47,7 @@ def error_no_backtrace()
 #[test]
 def error_bad_type()
   let type_err = try
-    diagnostic.render_error "boom" backtrace: "bad"
+    term.render_error "boom" backtrace: "bad"
   catch e
     e
   assert_eq (str type_err) "type error: expected strand.Backtrace"
